# simplify dd and let it use go runtime and kernel intelligence