With this change, dd runs in a single process, and uses zero copy when possible.
    Before this change, it ran at best 1/2 the speed of the GNU dd.
    WIth this change, it frequently is much faster.

```
    time ./dd if=/dev/zero of=/dev/null ibs=1048576 obs=1048576 status=progress count=65536
    68719476736 bytes (68719.477 MB, 65536.000 MiB) copied, 1.987 s, 34589.234 MB/s

    GNU dd
    time dd if=/dev/zero of=/dev/null ibs=1048576 obs=1048576 status=progress count=65536
    55726571520 bytes (56 GB, 52 GiB) copied, 3 s, 18.6 GB/s
    65536+0 records in
    65536+0 records out
    68719476736 bytes (69 GB, 64 GiB) copied, 3.69293 s, 18.6 GB/s
```

    the u-root dd is twice as fast for lots of data.

    For small data, it is still competitive:
```
    time dd if=4m of=/dev/null ibs=1048576 obs=1048576 status=progress count=65536
    4096+0 records in
    4096+0 records out
    4294967296 bytes (4.3 GB, 4.0 GiB) copied, 0.538755 s, 8.0 GB/s

    real    0m0.541s
    user    0m0.208s
    sys     0m0.333s
    rminnich@pop-os:~/go/src/github.com/u-root/u-root/cmds/core/dd$ time ./dd if=4m of=/dev/null ibs=1048576 obs=1048576 status=progress count=65536
    4294967296 bytes (4294.967 MB, 4096.000 MiB) copied, 0.449 s, 9571.651 MB/s

    real    0m0.453s
    user    0m0.049s
    sys     0m0.412s
```
    Before these changes, it was at best 50% as fast.